pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = ""
authors = [
    {name = "“2813204478@qq.com”",email = "“2813204478@qq.com”"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai (>=2.11.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "fastapi (>=0.124.4,<0.125.0)",
    "uvicorn (>=0.38.0,<0.39.0)"
]

[tool.poetry]
packages = [{include = "ai", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
